Issue #2747849: Add test coverage for timestamp/datetime conversion in addIndexField
